[feat] : 팜클럽 성공 바텀시트 디자인 변경 (#107)
